Extract common sources for Linux and MacOSX
